% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{browse_api_docs}
\alias{browse_api_docs}
\title{Browse the ProdPad API docs at swaggerhub.}
\usage{
browse_api_docs(version = "1.0")
}
\arguments{
\item{version}{Version number}
}
\description{
Browse the ProdPad API docs at swaggerhub.
}
