% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_prodpad_api.R
\name{describe}
\alias{describe}
\title{Loads prodpad api from internal data, then construct roxygen and code}
\usage{
describe(path = "/feedbacks", verb = "get", element = "parameters")
}
\description{
Loads prodpad api from internal data, then construct roxygen and code
}
\examples{
describe("/feedbacks")
describe("/ideas")
describe("/ideas/{id}")
}
\keyword{Internal}
