% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_prodpad_api.R
\name{describe_api}
\alias{describe_api}
\title{Describe prodpad API.}
\usage{
describe_api(max.level = 4)
}
\arguments{
\item{max.level}{passed to [str()]}
}
\description{
Describe prodpad API.
}
\examples{
describe_api()
}
\keyword{Internal}
