% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_prodpad_api.R
\name{describe_api_paths}
\alias{describe_api_paths}
\title{Describe prodpad API paths.}
\usage{
describe_api_paths()
}
\description{
Describe prodpad API paths.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
prodpad:::describe_api_paths()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
