% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_prodpad_date}
\alias{parse_prodpad_date}
\title{Parse prodpad dates using lubridate.}
\usage{
parse_prodpad_date(x)
}
\arguments{
\item{x}{Character vector to parse.  Assumes \code{datetimes} are written as \verb{2022-01-01T20-00-00Z}}
}
\description{
This is useful for columns such as \code{created_at} and \code{updated_at}.
}
