% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_companies}
\alias{pp_get_companies}
\title{Get a tibble of companies.}
\usage{
pp_get_companies(
  name = NULL,
  country = NULL,
  company_size = NULL,
  value = NULL,
  city = NULL,
  tags = NULL,
  external_id = NULL,
  external_url = NULL,
  contacts = FALSE,
  feedbacks = FALSE,
  page = 1,
  size = 100,
  .limit = size,
  .debug = FALSE,
  ...
)
}
\arguments{
\item{name}{Filter the companies by the name or partial name of the companies.}

\item{country}{Set to filter the companies based on the country. Use ISO Alpha-2 country codes. Only one country can be filtered at a time.}

\item{company_size}{Set to filter the companies based on their size.}

\item{value}{Set to filter the companies based on their value.}

\item{city}{Set to filter the companies based on city.}

\item{tags}{Filter companies by the tags associated to the feedback. Mulitple tags can be specified and acts as an OR. Use the tag ID or UUID.}

\item{external_id}{Filter the companies by an ID from a 3rd party application associated to the companies in Prodpad}

\item{external_url}{Filter the companies by a URL that is associated to a company.}

\item{contacts}{Whether to include contacts associated with each company in the results. Default is true.}

\item{feedbacks}{Whether to include the feedback for each contact associated to a company in the results. Default is true. Note if this is set to true then contacts will be returned whether or not contacts is set to true or false.}

\item{page}{The page of results to return (size is always 100).}

\item{size}{The number of results per page}

\item{.limit}{The maximum number of elements to return.}

\item{...}{Other arguments passed to \url{.pp()}]}
}
\description{
Return a list of companies. Companies are a collection of contacts that are providing feedback. Using companies allows you to group feedback from the same organisational source.
}
\details{
While very useful for business to busines,s companies can be used to create a any sort grouping of contacts for example by a conference or a cohort.

Using parameters allows you to also select whether to return feedback associated to the company via the contacts connected to the company. This is only available to accounts with an Advanced+ subscription.
}
\note{
\code{GET /companies}
}
\seealso{
Other Companies: 
\code{\link{pp_get_companies_vector}()}
}
\concept{Companies}
