% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_contact}
\alias{pp_get_contact}
\title{Get a contact.}
\usage{
pp_get_contact(id = NULL, feedbacks = NULL, ...)
}
\arguments{
\item{id}{Contact ID to fetch.}

\item{feedbacks}{Whether to include the feedback provided by the contact in the response or not.}

\item{...}{Other arguments passed to \code{\link[=pp]{pp()}}}
}
\description{
Returns the details of a contact. This will include a contacts PII (such as name and email) so be aware of the privacy of the individual when using this endpoint.
}
\details{
You can choose to include the feedback provided by the contact or not.
}
\note{
GET /contacts/{id}
}
