% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{pp_get_contacts}
\alias{pp_get_contacts}
\title{Get a list of all the contacts in the account.}
\usage{
pp_get_contacts(
  company = NULL,
  persona = NULL,
  job_role = NULL,
  name = NULL,
  email = NULL,
  ...
)
}
\arguments{
\item{company}{UUID of a company to filter contacts by}

\item{persona}{ID of a persona to filter contacts by}

\item{job_role}{UUID of a job role to filter contacts by.}

\item{name}{Name of contact or partial name of contacts to filter the list by}

\item{email}{Filter the contacts by an email}

\item{...}{Other arguments passed to [pp()]}
}
\description{
Get a list of all the contacts in the account.
}
