% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_contacts_vector}
\alias{pp_get_contacts_vector}
\title{Get named vector of contacts}
\usage{
pp_get_contacts_vector()
}
\value{
Named vector
}
\description{
Get named vector of contacts
}
\seealso{
Other Contacts: 
\code{\link{pp_get_contacts}()}
}
\concept{Contacts}
