% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{pp_get_feedback}
\alias{pp_get_feedback}
\title{Get a piece of feedback.}
\usage{
pp_get_feedback(id = NULL, ...)
}
\arguments{
\item{id}{Feedback ID.}

\item{...}{Other arguments passed to \code{\link[=pp]{pp()}}}
}
\description{
Return an individual piece of feedback. The feedback is returned with the details of the contact that provided the piece of feedback.
}
\note{
GET /feedbacks/{id}
}
