% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_idea_feedback}
\alias{pp_get_idea_feedback}
\title{}
\usage{
pp_get_idea_feedback(id, ...)
}
\arguments{
\item{id}{Numeric ID of the idea.}

\item{...}{Other arguments passed to [pp()]}
}
\description{
return(rawdat)
Get feedback associated to an idea.
}
\details{
Returns a tibble of all the feedback that is associated to the idea. This allows you to display the feedback in other apps for a particular idea or sync feedback between ProdPad and other applications.
}
\note{
GET /ideas/{id}/feedback
}
