% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_ideas}
\alias{pp_get_ideas}
\title{Get list of ideas.}
\usage{
pp_get_ideas(
  tags = NULL,
  product = NULL,
  persona = NULL,
  status = NULL,
  state = c("all", "active", "active public", "archived", "unsorted"),
  page = 1,
  size = 100,
  .limit = 100,
  ...
)
}
\arguments{
\item{tags}{One or more tag names to filter the ideas by. These act as an OR
not AND}

\item{product}{Name of a product to filter the ideas by.}

\item{persona}{Name of a persona to filter the ideas by.}

\item{status}{Name of a workflow status to filter the ideas by.}

\item{state}{Filters the returned ideas based on their state. Active Public
is same as portal in the UI. If not included then the endpoint returns all
active and active_public the same as the UI. "all" returns all unsorted and
active ideas.}

\item{size}{The number of results per page}

\item{...}{Other arguments passed to \url{.pp()}]}
}
\description{
Get list of ideas.
}
\seealso{
Other Ideas: 
\code{\link{pp_get_idea_feedback}()}
}
\concept{Ideas}
