% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_personas}
\alias{pp_get_personas}
\title{Get tibble of personas.}
\usage{
pp_get_personas(...)
}
\arguments{
\item{...}{Other arguments passed to \url{.pp()}]}
}
\description{
Get tibble of personas.
}
\note{
GET /personas
}
\seealso{
Other Persona: 
\code{\link{pp_get_personas_vector}()}
}
\concept{Persona}
