% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_personas_vector}
\alias{pp_get_personas_vector}
\title{Get named vector of personas.}
\usage{
pp_get_personas_vector()
}
\value{
Named vector
}
\description{
Get named vector of personas.
}
\seealso{
Other Persona: 
\code{\link{pp_get_personas}()}
}
\concept{Persona}
