% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_user}
\alias{pp_get_user}
\title{Get a user.}
\usage{
pp_get_user(id = NULL, ...)
}
\arguments{
\item{id}{Numeric ID of the user.}

\item{...}{Other arguments passed to \code{\link[=pp]{pp()}}}
}
\description{
Return the details on the user and their role.
}
\note{
GET /user/{id}
}
\seealso{
Other User: 
\code{\link{pp_get_users}()}
}
\concept{User}
