% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_users}
\alias{pp_get_users}
\title{Get a list of users.}
\usage{
pp_get_users(...)
}
\arguments{
\item{...}{Other arguments passed to \url{.pp()}]}
}
\value{
Tibble with user information
}
\description{
This endpoint returns a list of users with roles in the account.
}
\note{
\code{GET /users}
}
\seealso{
Other User: 
\code{\link{pp_get_user}()}
}
\concept{User}
