% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{pp_post_feedback}
\alias{pp_post_feedback}
\title{Create a new feedback.}
\usage{
pp_post_feedback(
  contact_id,
  contact_name = NULL,
  product_id,
  product_name = NA,
  feedback,
  ...
)
}
\arguments{
\item{contact_id}{\verb{string($uuid)} ID of the contact providing the feedback.
Either Contact ID or Contact name is required.}

\item{contact_name}{\code{string} Name of the contact providing the feedback. Either
Contact ID or Contact name is required.}

\item{product_id}{Product id. See \code{\link[=pp_get_products]{pp_get_products()}}}

\item{product_name}{Product name. See \code{\link[=pp_get_products]{pp_get_products()}}}

\item{feedback}{\code{string} (Required) The feedback. This field accepts HTML and
is stored as UTF-8.}

\item{...}{Other arguments passed to \url{.pp()}]}
}
\description{
Create a new feedback in the account. The feedback has to have either a name
or an email and the actual feedback. The email is used to add the feedback to
an existing contact.
}
\details{
Fields to create a feedback. Note either the contact ID or contact name is
required. Otherwise a 400 error will be returned. If a name and email field
are used, the value in the email field will be used to see if the contact
already exists and if it does the feedback will be automatically added to the
existing contact.
}
\note{
\code{POST /feedbacks}
}
