% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_prodpad_api.R
\name{describe_api_paths}
\alias{describe_api_paths}
\title{Describe prodpad API paths.}
\usage{
describe_api_paths()
}
\description{
Describe prodpad API paths.
}
\examples{
describe_api_paths()
}
\keyword{Internal}
