% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R, R/get.R
\name{pp_get_contacts}
\alias{pp_get_contacts}
\title{Get a list of all the contacts in the account.}
\usage{
pp_get_contacts(
  name = NULL,
  email = NULL,
  company = NULL,
  persona = NULL,
  job_role = NULL,
  tags = NULL,
  external_id = NULL,
  external_url = NULL,
  feedbacks = NULL,
  page = NULL,
  size = NULL,
  ...
)

pp_get_contacts(
  name = NULL,
  email = NULL,
  company = NULL,
  persona = NULL,
  job_role = NULL,
  tags = NULL,
  external_id = NULL,
  external_url = NULL,
  feedbacks = NULL,
  page = NULL,
  size = NULL,
  ...
)
}
\arguments{
\item{name}{Name of contact or partial name of contacts to filter the list by}

\item{email}{Filter the contacts by an email.}

\item{company}{UUID of a company to filter contacts by.}

\item{persona}{ID of a persona to filter contacts by.}

\item{job_role}{UUID of a job role to filter contacts by.}

\item{tags}{ID, UUID or name of one or more tags to filter the contacts by.}

\item{external_id}{Filter the contacts by an ID from another application such as a CRM}

\item{external_url}{Filter the contacts by a URL from another application such as a CRM}

\item{feedbacks}{Wheter to include the feedback for each contact in the results. Default is false.}

\item{page}{Page number of results to return.}

\item{size}{Number of results per page.}

\item{...}{Other arguments passed to \url{.pp()}]}
}
\description{
Get a tibble of all the contacts in the account. This can be used to sync contacts between ProdPad and CRMs.
}
\note{
GET /contacts
}
\seealso{
Other Contacts: 
\code{\link{pp_get_contacts_vector}()}
}
\concept{Contacts}
