% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{pp_get_feedback}
\alias{pp_get_feedback}
\title{Get tibble of feedback.}
\usage{
pp_get_feedback(
  state = c("active", "archived", "unsorted", "all"),
  product = NULL,
  customer = NULL,
  company = NULL,
  ...
)
}
\arguments{
\item{state}{Set to value of active for active, archived for archived feedback, unsorted for unsorted feedback and all for all feedback. Default is active}

\item{product}{Filter results by the product associated to feedback. Can either be the product UUID or product ID}

\item{customer}{Can be either numeric ID, UUID of a contact or contact email}

\item{company}{Set to filter the feedback results based on whether the feedback was entered for a contact linked to the company.}

\item{...}{Other arguments passed to [pp()]}
}
\description{
Get tibble of feedback.
}
