% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{pp_get_feedbacks}
\alias{pp_get_feedbacks}
\title{Get tibble of feedback.}
\usage{
pp_get_feedbacks(
  state = c("active", "archived", "unsorted", "all"),
  company = NULL,
  company_country = NULL,
  company_size = NULL,
  company_value = NULL,
  customer = NULL,
  product = NULL,
  persona = NULL,
  job_role = NULL,
  tags = NULL,
  has_ideas = NULL,
  external_id = NULL,
  external_url = NULL,
  group_by = FALSE,
  page = 1,
  size = 100,
  ...
)
}
\arguments{
\item{state}{Set to value of active for active, archived for archived feedback, unsorted for unsorted feedback and all for all feedback. Default is active.}

\item{company}{Set to filter the feedback results based on whether the feedback was entered for a contact linked to the company.}

\item{company_country}{Set to filter the feedback results based on the country set for the associated company.}

\item{company_size}{Set to filter the feedback results based on the size of the company for the associated company.}

\item{company_value}{Set to filter the feedback results based on the value of the company for the company assocaited to the feedback.}

\item{customer}{Can be either numeric ID, UUID of a contac or contact email.}

\item{product}{Filter results by the product associated to feedback. Can either be the product UUID or product ID.}

\item{persona}{Filter results by the persona associated to feedback. Can either be the persona UUID or persona ID.}

\item{job_role}{Filter results by the job role of the contact associated to the feedback. Use the JobRole UUID.}

\item{tags}{Filter feedback results by the tags associated to the feedback. Mulitple tags can be specified and acts as an OR. Use the tag ID or UUID.}

\item{has_ideas}{Whether the feedback is associated to one or more ideas. Default is either.}

\item{external_id}{Filter feedback to return the feedback associated with a specific External ID. An example of an external ID is the ID of a record in a CRM or ID of a ticket in a customer support application.}

\item{external_url}{Filter feedback to return the feedback associated with a specific external url. An example of an external url is that of a record in a CRM or a ticket in a customer support application}

\item{group_by}{Setting to customer will group the feedback by customer}

\item{page}{Set to page through the results. Default is 1.}

\item{size}{For setting the number of results per page. Default is 100.}

\item{...}{Other arguments passed to [pp()]}
}
\description{
Get tibble of feedback.
}
\note{
GET /feedbacks
}
