% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{pp_get_ideas}
\alias{pp_get_ideas}
\title{Get list of ideas.}
\usage{
pp_get_ideas(
  tags = NULL,
  product = NULL,
  persona = NULL,
  status = NULL,
  state = c("active", "active_public", "archived"),
  size = 20,
  ...
)
}
\arguments{
\item{tags}{One or more tag names to filter the ideas by. These act as an OR not AND}

\item{product}{Name of a product to filter the ideas by.}

\item{persona}{Name of a persona to filter the ideas by.}

\item{status}{Name of a workflow status to filter the ideas by.}

\item{state}{Filters the returned ideas based on their state. Active Public is same as portal in the UI. If not included then the endpoint returns all active and active_public the same as the UI.}

\item{size}{The number of results per page}

\item{...}{Other arguments passed to [pp()]}
}
\description{
Get list of ideas.
}
