% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{pp_post_feedback}
\alias{pp_post_feedback}
\title{Create a new feedback.}
\usage{
pp_post_feedback(
  contact_id,
  contact_name = NULL,
  product_id,
  product_name = NA,
  feedback,
  company_id = NULL,
  ...
)
}
\arguments{
\item{contact_id}{\verb{string($uuid)} ID of the contact providing the feedback.
Either Contact ID or Contact name is required.}

\item{feedback}{\code{string} (Required) The feedback. This field accepts HTML and
is stored as UTF-8.}

\item{company_id}{\verb{string($uuid)} UUID of the company to link the contact to.
The UUID can be determined using /GET companies endpoint.}

\item{...}{Other arguments passed to \code{\link[=pp]{pp()}}}

\item{name}{\code{string} Name of the contact providing the feedback. Either
Contact ID or Contact name is required.}

\item{email}{\code{string} The email of the contact. This is used to avoid
duplication of contacts. This can be any unique ID for each contact.}

\item{about}{\code{string} Text field about the contact. This will overright the
existing about if the contact already exists. This field accepts HTML and
is stored as UTF-8.}
}
\description{
Create a new feedback in the account. The feedback has to have either a name
or an email and the actual feedback. The email is used to add the feedback to
an existing contact.
}
\details{
Fields to create a feedback. Note either the contact ID or contact name is
required. Otherwise a 400 error will be returned. If a name and email field
are used, the value in the email field will be used to see if the contact
already exists and if it does the feedback will be automatically added to the
existing contact.
}
\note{
POST /feedbacks
}
