% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodpad-package.R
\docType{package}
\name{prodpad-package}
\alias{prodpad}
\alias{prodpad-package}
\title{prodpad: R Client for the ProdPad API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provide a low level wrapper `pp()` as well as higher level wrappers around the ProdPad API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrie.github.io/prodpad/}
  \item \url{https://github.com/andrie/prodpad}
  \item Report bugs at \url{https://github.com/andrie/prodpad/issues}
}

}
\author{
\strong{Maintainer}: Andrie de Vries \email{andrie@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
